git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/configurable-code-block-analysis-fixed-merges@39613 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/octagon/values/OctagonDoubleValue.java
@@ -50,7 +50,7 @@ public class OctagonDoubleValue extends OctagonNumericValue<Double> {
 
   @Override
   public OctagonNumericValue add(OctagonNumericValue pVal) {
-    return pVal.add(value);
+    return pVal.add(value.doubleValue());
   }
 
   @Override
