+++ src/org/sosy_lab/cpachecker/cpa/validvars/ValidVars.java
@@ -176,7 +176,7 @@ public class ValidVars implements Serializable {
         builder.put(functionName, entry.getValue());
       } else {
         if (entry.getValue() > 1) {
-          builder.put(functionName, (byte) (entry.getValue() - 1));
+          builder.put(functionName, (byte) (entry.getValue().byteValue() - 1));
         }
       }
     }
@@ -190,7 +190,7 @@ public class ValidVars implements Serializable {
       if (!functionName.equals(pFunctionName)) {
         builder.put(functionName, entry.getValue());
       } else {
-        builder.put(functionName, (byte) (entry.getValue() + 1));
+        builder.put(functionName, (byte) (entry.getValue().byteValue() + 1));
       }
     }
     if (!numFunctionCalled.containsKey(pFunctionName)) {
